REFACTORED login into a folder for better comprehension
