+++ frontend/src/components/Login/subcomponents/Backdrop.jsx
@@ -1,0 +1,9 @@
+import React from 'react';
+
+function Backdrop() {
+  return (
+    <div className="cyber-grid"></div>
+  );
+}
+
+export default Backdrop;
